Add comments and change the time for generating a block in 3 min
For the mining process, we may be  to need change some parameter such as the time to generate a block, and  in this test, we change it to 3 min, and reduce the difficult to be generate block difficult mul 0.3 after it. Hope it work for generating a block in 3 mins.