Add a command for client
Add senrawtransactionabe for testing, when the wallet creates a transaction, it wll send the raw transaction to abec, abec will handle with this transaction, if valid, this transaction will be added into mempool, and relay to other peers. if not valid, a error will be sent back to the client of wallet.